skip manipulating payment/googlecheckout for recent magento versions
Mage_Sales does not have a dependency on Mage_GoogleCheckout anymore, at least not since 1.9.0.0, possibly earlier.

the workaround used has some side effects for extensions that make use of Mage::helper('sales')->getPaymentMethods(), as magento-mock only ads the code, "active" and "model". in our case, the (disabled) payment method was included in the configuration setting with <source_model>adminhtml/system_config_source_payment_allmethods</source_model> - as the title is not shown, it's listed at the top with an empty entry missleading the user.