Revert "Add mock for Mage_Log_Helper_Data"
This reverts commit f05658c9cf55a06b64cfde6118ad51d96e2bd1df.

The Mage Log helper is loaded correctly even when the Mage_Log module
is disabled. This is due to the way Magento falls back to "Mage_<module>"
when it cant find any reference in configuration for a given rewrite
style helper name.

Adding the mock means the log filename validation that was added as part
of the SUPEE-10415 security patch is removed. Reverting this commit
ensures we maintain all the functionality introduced in the security
patch.