Update the modman file to not use a wildcard
The modman file references a wildcard for the modules xml, which in some
versions of the Magento composer installer would result in a gitignore
line being added for the entire app/etc/modules directory. The file has
been updated to explicitly reference the Zookal_Mock.xml file to ensure
that this never happens.